Trigger new events in shift and select calls
Added text_input_{}_select event when a character is selected in the widget. The event contains the current value of the text widget and length for handlers to use. Added text_input_{}_active_character event when the active character is cycled. The event contains the newly active character for handlers to use.