Group Negative Digits
This changes from isDigit (only finds positive numbers) to casting to int to find all numbers. If an int is found, it converts it back to a string and strips the negative sign from it.  It will then group as normal and replace the number portion with the grouped digits and leave the negative sign intact.